Separate model and validate checks in metric system model
